Add files via upload
(1) [Processed dataset]: It is available with the article in the supplemental material at "Appendix A". The file is now available in this repository as file name "Jas et al. 2024_Supply_EML_Gravel_Liq_COGE.pdf". This is the same dataset used for the model training and testing. This same file is attached to the COMGE journal section.