Add files via upload
These are the LightGBM model files (Initial Trial) with two different extensions. This file can be loaded in the Python platform to predict the liquefaction potential assessment of gravelly soils for site(s) of interest having all IPs in a excel spreadsheet.
The final/best trial will be uploaded soon.